refactor(storage): extract the entity subsystem into storage/entity-store.ts
God-file decomposition (#1526 playbook applied to StorageManager).
Moves the entity file subsystem verbatim (604 LOC, 8 methods):
normalizeEntityName, writeEntity (schema-validated structured
sections), readEntity, listEntityNames, updateEntitySynthesis,
readAllEntityFiles, buildKnowledgeIndex, mergeFragmentedEntities —
into EntityStore (18 deps incl. the ctor parameter-properties and the
entitiesDir getter, all live via selfDeps). Entity helpers exported
through the ../index.js barrel so directStorageImports stays 42;
KNOWLEDGE_INDEX_CACHE_TTL_MS static widened from private.

storage.ts: 7,747 -> 7,227 LOC.
